fin-projet

+++ app/Http/Controllers/TeamController.php
@@ -110,7 +110,7 @@ class TeamController extends Controller
      */
     public function destroy(Team $team)
     {
-        {   if (Storage::exists($team->photo)) {
+          if (Storage::exists($team->photo)) {
             Storage::disk('public')->delete($team->photo);
         } 
         $team->delete();
